+++ commands/music/play.js
@@ -31,7 +31,7 @@ module.exports.run = async (client,message,args,prefix,player) => {
         return void message.reply("Could not join your voice channel!");
     }
 
-    await message.reply(`⏱ | Loading your ${searchResult.playlist ? "playlist" : "track"}...`);
+    await message.reply(`⏱ | Загружаю твой ${searchResult.playlist ? "плейлист" : "трек"}...`);
     searchResult.playlist ? queue.addTracks(searchResult.tracks) : queue.addTrack(searchResult.tracks[0]);
     if (!queue.playing) await queue.play();
 };